+++ index.js
@@ -68,10 +68,16 @@ module.exports = {
 			'all'
 		],
 
-		// https://www.npmjs.com/package/eslint-plugin-dependencies
+		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciescase-sensitive
 		'dependencies/case-sensitive': 1,
+
+		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesno-cycles
 		'dependencies/no-cycles': 1,
+
+		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesno-unresolved
 		'dependencies/no-unresolved': 1,
+
+		// https://www.npmjs.com/package/eslint-plugin-dependencies#dependenciesrequire-json-ext
 		'dependencies/require-json-ext': 1,
 
 		// http://eslint.org/docs/rules/dot-notation
@@ -226,6 +232,18 @@ module.exports = {
 			}
 		],
 
+		// https://github.com/nodejs/node/blob/8191af5b292aa5d5f07492105781b6cf1d91c42f/.eslintrc.yaml#L104
+		'no-restricted-syntax': [2, {
+			selector: "CallExpression[callee.name='setTimeout'][arguments.length<2]",
+			message: "setTimeout() must be invoked with at least two arguments."
+		}, {
+			selector: "CallExpression[callee.name='setInterval'][arguments.length<2]",
+			message: "setInterval() must be invoked with at least 2 arguments."
+		}, {
+			selector: "ThrowStatement > CallExpression[callee.name=/Error$/]",
+			message: "Use new keyword when throwing an Error."
+		}],
+
 		// http://eslint.org/docs/rules/no-self-assign
 		'no-self-assign': [
 			'error',
@@ -233,6 +251,9 @@ module.exports = {
 				props: true
 			}
 		],
+
+		// https://eslint.org/docs/rules/no-throw-literal
+		'no-throw-literal': 'error',
 
 		// http://eslint.org/docs/rules/no-var
 		// https://github.com/airbnb/javascript#references--disallow-var